Main Analysis
This is an Excel file where I performed most of my analysis gathering data showing what a student might pay in Oregon, and what that same student might pay elsewhere. There are links to the data sources in the case someone wanted to replicate the work. It is basically a loose comparison between the tuition and basic living in Oregon vs US and Canada, England, Germany and France.